suggest: Move Remote Settings-related types into a separate module.
This commit moves the types and updates their visibility, without any
logic changes.
